feat: it works
- Updated session configuration parameters for calibration and detection thresholds.
- Enhanced WebSocket touch event handling to preserve event types and added logging for received messages.
- Modified TouchEvent class to ensure position values are floats.
- Improved marker detection to support variable grid sizes and added debugging information.
- Refactored marker projector to handle a flexible number of projector points and added margin validation.
- Updated homography computation to support N-point correspondences and improved error handling.
- Adjusted calibration logic to validate and utilize a grid of calibration points instead of fixed corners.
- Enhanced CalibrationResult validation to require at least 4 points for depth and RGB corners.

+++ code/cv-system/src/cv_system/bridge/touch_event.py
@@ -62,7 +62,7 @@ class TouchEvent:
 
         self.type = event_type
         self.touch_id = touch_id
-        self.position = {"x": round(x, 2), "y": round(y, 2)}
+        self.position = {"x": float(round(x, 2)), "y": float(round(y, 2))}
         self.timestamp = timestamp
 
         logger.debug(
